shared session check
